Lots of things are a mess right now sorry
Was having an issue with the game mode not capturing the mouse when clicking on the game, leading to issues with controls and having to hold right click to turn camera. Copying stuff over from the third person starter content to fix the issue since it was not apparent at all as to how to fix.