Rifle pickup added to game with hitscan detection
Rifle can be picked up on BigFlat with f, works in multiplayer. However, I am not currently able to view my own server on Steam so that function may be broken. None of those files have been changed since it was previously working, so I am unsure of the cause.